# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Testing Swagger APIs.",
        "title": "Go RESTful API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/users": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Description...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get users.",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/server.HandleGetUsers.response"
                        }
                    },
                    "500": {
                        "description": "Error Internal Server",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Description...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create user.",
                "parameters": [
                    {
                        "description": "User params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.HandleCreateUser.request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/server.HandleCreateUser.response"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Error Internal Server",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        },
        "/users/{email}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Description...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by email.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email user",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/server.HandleGetUserByEmail.response"
                        }
                    },
                    "500": {
                        "description": "Error Internal Server",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.HandleCreateUser.request": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 25
                },
                "email": {
                    "type": "string",
                    "example": "alonso12.dev@gmail.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "Alonso"
                },
                "last_name": {
                    "type": "string",
                    "example": "Hdez"
                },
                "name": {
                    "type": "string",
                    "example": "Jorge Luis"
                },
                "phone": {
                    "type": "string",
                    "example": "7713037204"
                },
                "position": {
                    "type": "string",
                    "example": "Go developer"
                }
            }
        },
        "server.HandleCreateUser.response": {
            "type": "object",
            "properties": {
                "datetime": {
                    "type": "string"
                },
                "http_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/server.userAddResponse"
                }
            }
        },
        "server.HandleGetUserByEmail.response": {
            "type": "object",
            "properties": {
                "datetime": {
                    "type": "string"
                },
                "http_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/server.userGetResponse"
                }
            }
        },
        "server.HandleGetUsers.response": {
            "type": "object",
            "properties": {
                "datetime": {
                    "type": "string"
                },
                "http_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "user": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.userGetResponse"
                    }
                }
            }
        },
        "server.errorResponse": {
            "type": "object",
            "properties": {
                "datetime": {
                    "type": "string"
                },
                "details": {
                    "type": "object"
                },
                "error": {
                    "type": "string"
                },
                "http_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "server.userAddResponse": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "server.userGetResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}